Allow Span.TryCopyTo to be inlined into caller - Have single return from TryCopyTo - Move 'nuint' parameters up the call stack for more efficient use of registers

+++ src/mscorlib/shared/System/Span.NonGeneric.cs
@@ -242,7 +242,7 @@ namespace System
             return new ReadOnlySpan<char>(ref Unsafe.Add(ref text.GetRawStringData(), start), length);
         }
 
-        internal static unsafe void CopyTo<T>(ref T destination, ref T source, int elementsCount)
+        internal static unsafe void CopyTo<T>(ref T destination, ref T source, nuint elementsCount)
         {
             if (Unsafe.AreSame(ref destination, ref source))
                 return;
@@ -256,7 +256,7 @@ namespace System
                 return;
             }
 
-            nuint byteCount = (nuint)elementsCount * (nuint)Unsafe.SizeOf<T>();
+            nuint byteCount = elementsCount * (nuint)Unsafe.SizeOf<T>();
             if (!RuntimeHelpers.IsReferenceOrContainsReferences<T>())
             {
                 Buffer.Memmove(
